requirements.txt
attrs==19.3.0
backcall==0.1.0
bezier==2020.5.19
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
cvxopt==1.2.5
cvxpy==1.1.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
ecos==2.0.7.post1
entrypoints==0.3
ForceAtlas2==1.0
future==0.18.2
hdbscan==0.8.26
imageio==2.9.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.1
olefile==0.46
osqp==0.6.2
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycparser==2.20
pygame==1.9.6
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyzmq==18.1.1
qdldl==0.1.5.post0
qtconsole==4.7.3
QtPy==1.9.0
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.4.1
scs==2.1.2
seaborn==0.11.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.1.8
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
